Generic slider cherry pick Make slider generic Add slider with type double example Move some shared input helper functions to static class Add description to ApiDocumentation
