+++ Directory.Build.props
@@ -10,7 +10,7 @@
 	 <RepositoryType>git</RepositoryType>
 	 <PublishRepositoryUrl>true</PublishRepositoryUrl>
 	 <SolutionDir>$(MSBuildThisFileDirectory)</SolutionDir>
-	 <VersionPrefix>2.3.7</VersionPrefix>
+	 <VersionPrefix>2.4.0</VersionPrefix>
 	 <VersionSuffix></VersionSuffix> 
 	 <AccelerateBuildsInVisualStudio>true</AccelerateBuildsInVisualStudio>
  </PropertyGroup>
