+++ examples/FluentUI.Demo.Shared/Pages/NumberField/NumberFieldPage.razor
@@ -12,7 +12,19 @@
 </p>
 
 
-<ApiDocumentation Component="typeof(FluentNumberField<>)" InstanceType="typeof(int)" GenericLabel="TValue" />
+<ApiDocumentation Component="typeof(FluentNumberField<>)" InstanceType="typeof(int)" GenericLabel="TValue">
+    <Description>
+        <code>TValue</code> can be one of the following <code>Type</code>'s:
+        <ul>
+            <li><code>int</code></li>
+            <li><code>long</code></li>
+            <li><code>short</code></li>
+            <li><code>float</code></li>
+            <li><code>double</code></li>
+            <li><code>decimal</code></li>
+        </ul>
+    </Description>
+</ApiDocumentation>
     
 <h2>Examples</h2>
