+++ examples/FluentUI.Demo.Shared/wwwroot/sources/SliderDouble.razor.txt
@@ -1,0 +1,13 @@
+﻿<div style="margin-bottom: 60px;">
+    <FluentSlider Min="-1.50" Max="7.0" Step="0.5" @bind-Value=value>
+        <FluentSliderLabel Position="-1">-1&#8451;</FluentSliderLabel>
+        <FluentSliderLabel Position="1.5">1.5&#8451;</FluentSliderLabel>
+        <FluentSliderLabel Position="5">5&#8451;</FluentSliderLabel>
+        <FluentSliderLabel Position="6">6&#8451;</FluentSliderLabel>
+    </FluentSlider>
+</div>
+<p>Selected value: @value</p>
+
+@code {
+    double value = 3.5;
+}
