+++ src/Microsoft.Fast.Components.FluentUI/Components/Slider/FluentSliderLabel.razor.cs
@@ -1,14 +1,15 @@
+using System.Globalization;
 using Microsoft.AspNetCore.Components;
 
 namespace Microsoft.Fast.Components.FluentUI;
 
-public partial class FluentSliderLabel : FluentComponentBase
+public partial class FluentSliderLabel<TValue> : FluentComponentBase
 {
     /// <summary>
-    /// Gets or sets the slider's label position
+    /// Gets or sets the value for this slider position.
     /// </summary>
     [Parameter]
-    public int? Position { get; set; }
+    public TValue? Position { get; set; }
 
     /// <summary>
     /// Gets or sets if marks are hidden
@@ -26,4 +27,20 @@ public partial class FluentSliderLabel : FluentComponentBase
     /// </summary>
     [Parameter]
     public RenderFragment? ChildContent { get; set; }
+
+    protected string? FormatValueAsString(TValue? value)
+    {
+        // Avoiding a cast to IFormattable to avoid boxing.
+        return value switch
+        {
+            null => null,
+            int @int => BindConverter.FormatValue(@int, CultureInfo.InvariantCulture),
+            long @long => BindConverter.FormatValue(@long, CultureInfo.InvariantCulture),
+            short @short => BindConverter.FormatValue(@short, CultureInfo.InvariantCulture),
+            float @float => BindConverter.FormatValue(@float, CultureInfo.InvariantCulture),
+            double @double => BindConverter.FormatValue(@double, CultureInfo.InvariantCulture),
+            decimal @decimal => BindConverter.FormatValue(@decimal, CultureInfo.InvariantCulture),
+            _ => throw new InvalidOperationException($"Unsupported type {value.GetType()}"),
+        };
+    }
 }
